requirements.txt
disnake
aiohttp
motor
disnake-jishaku
easy_pil
dnspython
disnake
aiopyston
Flask 
googletrans==4.0.0rc1